Combine navigation of 2 platforms into one and fix some UI bugs on iOS
