+++ gradle.properties
@@ -1,3 +1,4 @@
+kotlin.native.cacheKind=none
 #Gradle
 # Increase maximum heap size
 org.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=1g -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
@@ -14,5 +15,6 @@ android.nonTransitiveRClass=true
 kotlin.native.useXcodeEnv=true
 
 ksp.useKSP2=true
+org.gradle.unsafe.configuration-cache=false
 
 org.gradle.java.home=/Users/tu.nguyen/Library/Java/JavaVirtualMachines/corretto-21.0.7/Contents/Home
